Fix case sensitivity issue in Task 13 delete_a script

+++ python-object_relational_mapping/13-model_state_delete_a.py
@@ -17,7 +17,8 @@ if __name__ == "__main__":
 
     states = session.query(State).filter(State.name.like('%a%')).all()
     for state in states:
-        session.delete(state)
+        if 'a' in state.name:
+            session.delete(state)
     session.commit()
 
     session.close()
